Update Java version and simplify caching in CI workflow
Updated the application's Java version from 17 to 20 in both GitHub Action's gradle_ci.yml and build.gradle.kts files. This is to ensure the application is using the latest Java version and benefiting from the latest language features and security patches.

Also, simplified the caching mechanism in the gradle_ci.yml file. Replaced the usage of actions/cache with a simple 'cache' command-line argument, making the usage more concise and direct. Streamlined the condition for dropbox to avoid unnecessary line breaks. This could help in improving the readability and maintainability of the workflow file.